requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
dj-database-url==1.0.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-bootstrap-datepicker-plus==5.0.2
django-bootstrap4==22.3
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pydantic==1.10.2
pytz==2022.6
sqlparse==0.4.3
